transfer Chat-component to Chat.js
